[PC-225] MatchingDetail 파일 분리
