+++ feature/matching/src/main/java/com/puzzle/matching/detail/content/ProfileValuePickBody.kt
@@ -1,0 +1,75 @@
+package com.puzzle.matching.detail.content
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.material3.Tab
+import androidx.compose.material3.TabRow
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.puzzle.designsystem.foundation.PieceTheme
+import com.puzzle.domain.model.pick.ValuePick
+
+@Composable
+internal fun ProfileValuePickBody(
+    pickCards: List<ValuePick>,
+    modifier: Modifier = Modifier,
+) {
+    val tabIndex = remember { mutableIntStateOf(0) }
+
+    val tabTitles = listOf("전체", "나와 같은", "나와 다른")
+
+    Column(modifier = modifier.fillMaxSize()) {
+        TabRow(
+            selectedTabIndex = tabIndex.intValue,
+            modifier = Modifier.fillMaxWidth(),
+        ) {
+            tabTitles.forEachIndexed { index, title ->
+                Tab(
+                    selected = (tabIndex.intValue == index),
+                    onClick = { tabIndex.intValue = index },
+                    text = { Text(text = title) },
+                )
+            }
+        }
+
+        when (tabIndex.intValue) {
+            0 -> TabContent("전체 탭의 내용 예시...")
+            1 -> TabContent("나만 탭의 내용 예시...")
+            2 -> TabContent("너만 탭의 내용 예시...")
+        }
+    }
+}
+
+@Composable
+private fun TabContent(contentText: String) {
+    LazyColumn(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(16.dp),
+    ) {
+        items(15) { index ->
+            Text(
+                text = "$contentText 아이템 $index",
+                modifier = Modifier.padding(vertical = 8.dp),
+            )
+        }
+    }
+}
+
+@Preview
+@Composable
+private fun ProfileValuePickBodyPreview() {
+    PieceTheme {
+        ProfileValuePickBody(
+            pickCards = emptyList()
+        )
+    }
+}
